Fix BitView in Windows XP

+++ plugins/bitview_app/bit_container.cpp
@@ -128,7 +128,7 @@ namespace satdump
                         if (d_display_mode == 0) // Bit display
                         {
 #pragma omp parallel for
-                            for (int64_t line = 0; (size_t)line < d_chunk_size; line++)
+                            for (int64_t line = 0; line < d_chunk_size; line++)
                             {
                                 for (size_t i = 0; i < d_chunk_size; i++)
                                 {
@@ -161,7 +161,7 @@ namespace satdump
                         else if (d_display_mode == 1) // Byte display
                         {
 #pragma omp parallel for
-                            for (int64_t line = 0; (size_t)line < d_chunk_size; line++)
+                            for (int64_t line = 0; line < d_chunk_size; line++)
                             {
                                 for (size_t i = 0; i < d_chunk_size / 8; i++)
                                 {
